requirements.txt
mysql-connector-python
logging-py
enum-compat